+++ client/src/pages/Home/index.ts
@@ -1,1 +1,1 @@
-export * from "./Home";
+export * from "./Home_";
